refactor(vite-plugin-angular): move module resolution into host augmentation
